add & import Glyphicons
!IMPORTANT NOTE!

Due to Glyphicons library being proprietary,
the package is not available as open-source and cannot be
installed via package manager. Therefore I have manually added it to
`/web/themes/custom/legacy/styles/vendor/`,
and so you must update it manually when a new version of Glyphicons
is released.

The directory above is **gitignored**, because we don’t want
third-party libraries in our git history. (It is also very large—91MB.)
Therefore you must be sure to transfer the package via FTP
when building a new child site, and manually keep versions up-to-date.