rename `.h-Constrain` -> `.o-Constrain`
This project’s `.o-Constrain` Object is DEPRECATED,
in favor of Xmeter’s `.h-Constrain` Helper.

For details, see:
<https://chharvey.github.io/xmeter/docs/styleguide/section-helpers.html>

NOTE: This commit marks DEPRECATED, but does not remove,
the internal `.o-Constrain` Object.
This is because there is dynamic content that uses the class,
namely, the submenus in the header.

Once the menus are automated and markup does not need to be
manually entered, the `.o-Constrain` Object can be removed
from this project. To achieve this, do the following steps:
(all filepaths relative to `/web/themes/custom/legacy/`)

1. DELETE file `css/src/_o-Constrain.less`
2. from `css/src/legacy.less`:
  A. remove `@import (less) url('_o-Constrain.less');`
  B. remove corresponding entry in the table of contents
3. from `legacy.libraries.yml`, remove `css/dist/_o-Constrain.css`
  from `css/layout` section